+++ src/download_sources/fdroid.rs
@@ -13,7 +13,7 @@ use futures_util::StreamExt;
 use indicatif::MultiProgress;
 use regex::Regex;
 use ring::digest::{Context, SHA256};
-use serde_json::Value;
+use serde_json::{json, Value};
 use sha1::{Sha1, Digest as Sha1Digest};
 use sha2::Sha256;
 use simple_error::SimpleError;
@@ -24,15 +24,15 @@ use x509_certificate::certificate::CapturedX509Certificate;
 
 use crate::consts;
 use crate::config::{self, ConfigDirError};
-use crate::progress_bar::progress_wrapper;
+use crate::util::{OutputFormat, progress_bar::progress_wrapper};
 mod error;
 use error::Error as FDroidError;
 
-async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgress>) -> Value {
+async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgress>, output_format: OutputFormat) -> Value {
     let temp_dir = match tempdir() {
         Ok(temp_dir) => temp_dir,
         Err(_) => {
-            println!("Could not create temporary directory for F-Droid package index. Exiting.");
+            print_error("Could not create temporary directory for F-Droid package index. Exiting.", output_format);
             std::process::exit(1);
         }
     };
@@ -49,7 +49,7 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
             fingerprint = match hex::decode(fingerprint_option) {
                 Ok(hex_fingerprint) => hex_fingerprint,
                 Err(_) => {
-                    println!("Fingerprint must be specified as valid hex. Exiting.");
+                    print_error("Fingerprint must be specified as valid hex. Exiting.", output_format);
                     std::process::exit(1);
                 }
             };
@@ -62,10 +62,10 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
     let display_error_and_exit = |err: ConfigDirError| {
         match err {
             ConfigDirError::NotFound => {
-                println!("Could not find a config directory for apkeep to store F-Droid package index. Exiting.");
+                print_error("Could not find a config directory for apkeep to store F-Droid package index. Exiting.", output_format.clone());
             },
             ConfigDirError::CouldNotCreate => {
-                println!("Could not create a config directory for apkeep to store F-Droid package index. Exiting.");
+                print_error("Could not create a config directory for apkeep to store F-Droid package index. Exiting.", output_format.clone());
             },
         }
         std::process::exit(1);
@@ -90,7 +90,7 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
         Ok(mut file) => {
             let mut contents = String::new();
             if file.read_to_string(&mut contents).is_err() {
-                println!("Could not read etag file for F-Droid package index. Exiting.");
+                print_error("Could not read etag file for F-Droid package index. Exiting.", output_format);
                 std::process::exit(1);
             }
             Some(contents)
@@ -111,7 +111,7 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
     let etag = if jar_response.headers().contains_key("ETag") {
         jar_response.headers()["ETag"].to_str().unwrap()
     } else {
-        println!("Could not receive etag for F-Droid package index. Exiting.");
+        print_error("Could not receive etag for F-Droid package index. Exiting.", output_format);
         std::process::exit(1);
     };
 
@@ -125,20 +125,20 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
         let index = read_file_to_string(index_file);
         serde_json::from_str(&index).unwrap()
     } else {
-        let files = download_and_extract_to_tempdir(&temp_dir, &repo, Rc::clone(&mp), use_entry).await;
+        let files = download_and_extract_to_tempdir(&temp_dir, &repo, Rc::clone(&mp), use_entry, output_format.clone()).await;
         let verify_index = match options.get("verify-index") {
             Some(&"false") => false,
             _ => true,
         };
-        match verify_and_return_json(&temp_dir, &files, &fingerprint, verify_index, use_entry) {
+        match verify_and_return_json(&temp_dir, &files, &fingerprint, verify_index, use_entry, Rc::clone(&mp)) {
             Ok(json) => {
                 let index = if use_entry {
-                    match verify_and_return_index_from_entry(&temp_dir, &repo, &json, verify_index, mp).await {
+                    match verify_and_return_index_from_entry(&temp_dir, &repo, &json, verify_index, mp, output_format.clone()).await {
                         Ok(index_from_entry) => {
                             index_from_entry
                         }
                         Err(_) => {
-                            println!("Could verify and return package index from entry JSON. Exiting.");
+                            print_error("Could verify and return package index from entry JSON. Exiting.", output_format);
                             std::process::exit(1);
                         }
                     }
@@ -149,26 +149,33 @@ async fn retrieve_index_or_exit(options: &HashMap<&str, &str>, mp: Rc<MultiProgr
                 match serde_json::from_str(&index) {
                     Ok(index_value) => {
                         if fs::write(index_file, index).is_err() {
-                            println!("Could not write F-Droid package index to config file. Exiting.");
+                            print_error("Could not write F-Droid package index to config file. Exiting.", output_format);
                             std::process::exit(1);
                         }
                         if fs::write(latest_etag_file, etag).is_err() {
-                            println!("Could not write F-Droid etag to config file. Exiting.");
+                            print_error("Could not write F-Droid etag to config file. Exiting.", output_format);
                             std::process::exit(1);
                         }
                         index_value
                     }
                     Err(_) => {
-                        println!("Could not decode JSON for F-Droid package index. Exiting.");
+                        print_error("Could not decode JSON for F-Droid package index. Exiting.", output_format);
                         std::process::exit(1);
                     }
                 }
             },
             Err(_) => {
-                println!("Could not verify F-Droid package index. Exiting.");
+                print_error("Could not verify F-Droid package index. Exiting.", output_format);
                 std::process::exit(1);
             },
         }
+    }
+}
+
+fn print_error(err_msg: &str, output_format: OutputFormat) {
+    match output_format {
+        OutputFormat::Plaintext => eprintln!("{}", err_msg),
+        OutputFormat::Json => println!("{{\"error\":\"{}\"}}", err_msg),
     }
 }
 
@@ -187,10 +194,10 @@ pub async fn download_apps(
     options: HashMap<&str, &str>,
 ) {
     let mp = Rc::new(MultiProgress::new());
-    let mp_index = Rc::clone(&mp);
-    let index = retrieve_index_or_exit(&options, mp_index).await;
+    let index = retrieve_index_or_exit(&options, Rc::clone(&mp), OutputFormat::Plaintext).await;
 
-    let (fdroid_apps, repo_address) = match parse_json_for_download_information(index, apps) {
+    let app_arch = options.get("arch").map(|x| x.to_string());
+    let (fdroid_apps, repo_address) = match parse_json_for_download_information(index, apps, app_arch.clone(), Rc::clone(&mp)) {
         Ok((fdroid_apps, repo_address)) => (fdroid_apps, repo_address),
         Err(_) => {
             println!("Could not parse JSON of F-Droid package index. Exiting.");
@@ -205,15 +212,24 @@ pub async fn download_apps(
             let repo_address = Rc::clone(&repo_address);
             let mp_log = Rc::clone(&mp);
             let mp = Rc::clone(&mp);
+            let app_arch = app_arch.clone();
             async move {
-                let app_string = match app_version {
-                    Some(ref version) => {
-                        mp_log.println(format!("Downloading {} version {}...", app_id, version)).unwrap();
+                let app_string = match (app_version, app_arch) {
+                    (None, None) => {
+                        mp_log.suspend(|| println!("Downloading {}...", app_id));
+                        app_id.to_string()
+                    },
+                    (None, Some(arch)) => {
+                        mp_log.suspend(|| println!("Downloading {} arch {}...", app_id, arch));
+                        format!("{}@{}", app_id, arch)
+                    },
+                    (Some(version), None) => {
+                        mp_log.suspend(|| println!("Downloading {} version {}...", app_id, version));
                         format!("{}@{}", app_id, version)
                     },
-                    None => {
-                        mp_log.println(format!("Downloading {}...", app_id)).unwrap();
-                        app_id.to_string()
+                    (Some(version), Some(arch)) => {
+                        mp_log.suspend(|| println!("Downloading {} version {} arch {}...", app_id, version, arch));
+                        format!("{}@{}@{}", app_id, version, arch)
                     },
                 };
                 let fname = format!("{}.apk", app_string);
@@ -258,9 +274,9 @@ pub async fn download_apps(
                         };
                         if let Some(sha256sum) = sha256sum {
                             if sha256sum == hash {
-                                mp_log.println(format!("{} downloaded successfully!", app_string)).unwrap();
+                                mp_log.suspend(|| println!("{} downloaded successfully!", app_string));
                             } else {
-                                mp_log.println(format!("{} downloaded, but the sha256sum does not match the one signed by F-Droid. Proceed with caution.", app_string)).unwrap();
+                                mp_log.suspend(|| println!("{} downloaded, but the sha256sum does not match the one signed by F-Droid. Proceed with caution.", app_string));
                             }
                         }
                     },
@@ -278,7 +294,7 @@ type DownloadInformation = (Vec<(String, Option<String>, String, Vec<u8>)>, Stri
 /// flexible enough to parse either, and may work on future index versions as well.  Since `sha256`
 /// digests are checked before proceeding, I don't foresee this having an insecure failure mode, so
 /// checking the index version and making the parsing overly brittle has no substantive advantage.
-fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<String>)>) -> Result<DownloadInformation, FDroidError> {
+fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<String>)>, app_arch: Option<String>, mp_log: Rc<MultiProgress>) -> Result<DownloadInformation, FDroidError> {
     let index_map = index.as_object().ok_or(FDroidError::Dummy)?;
     let repo_address = index_map
         .get("repo").ok_or(FDroidError::Dummy)?
@@ -293,22 +309,25 @@ fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<S
         let (app_id, app_version) = app;
         match packages.get(&app_id) {
             Some(Value::Array(app_array)) => {
-                if app_version.is_none() {
-                    if !app_array.is_empty() && app_array[0].is_object() {
-                        let fdroid_app = app_array[0].as_object().unwrap();
-                        if let (Some(Value::String(filename)), Some(Value::String(hash))) = (fdroid_app.get("apkName"), fdroid_app.get("hash")) {
-                            if let Ok(hash) = hex::decode(hash.to_string()) {
-                                return Some((app_id, app_version, filename.to_string(), hash));
-                            }
-                        }
-                    }
-                } else {
-                    for single_app in app_array {
-                        if let Value::Object(fdroid_app) = single_app {
-                            if let Some(Value::String(version_name)) = fdroid_app.get("versionName") {
-                                if version_name == app_version.as_ref().unwrap() {
-                                    if let (Some(Value::String(filename)), Some(Value::String(hash))) = (fdroid_app.get("apkName"), fdroid_app.get("hash")) {
-                                        if let Ok(hash) = hex::decode(hash.to_string()) {
+                for single_app in app_array {
+                    if let Value::Object(fdroid_app) = single_app {
+                        if let Some(Value::String(version_name)) = fdroid_app.get("versionName") {
+                            if app_version.is_none() || version_name == app_version.as_ref().unwrap() {
+                                if let (Some(Value::String(filename)), Some(Value::String(hash))) = (fdroid_app.get("apkName"), fdroid_app.get("hash")) {
+                                    if let Ok(hash) = hex::decode(hash.to_string()) {
+                                        if let Some(arch) = &app_arch {
+                                            if let Some(Value::Array(nativecode_array)) = fdroid_app.get("nativecode") {
+                                                if nativecode_array.iter().any(|value| {
+                                                    if let Value::String(value_str) = value{
+                                                        value_str == arch
+                                                    } else {
+                                                        false
+                                                    }
+                                                }) {
+                                                    return Some((app_id, app_version, filename.to_string(), hash));
+                                                }
+                                            }
+                                        } else {
                                             return Some((app_id, app_version, filename.to_string(), hash));
                                         }
                                     }
@@ -316,9 +335,10 @@ fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<S
                             }
                         }
                     }
-                    println!("Could not find version {} of {}. Skipping...", app_version.unwrap(), app_id);
-                    return None;
                 }
+                let arch_str = app_arch.as_ref().map_or("".to_string(), |x| format!(" {}", x));
+                mp_log.println(format!("Could not find version {}{} of {}. Skipping...", app_version.unwrap(), arch_str, app_id)).unwrap();
+                return None;
             },
             Some(Value::Object(app_object)) => {
                 if let Some(Value::Object(versions)) = app_object.get("versions") {
@@ -359,7 +379,7 @@ fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<S
                     }
                 }
             },
-            _ => println!("Could not find {} in package list. Skipping...", app_id),
+            _ => mp_log.println(format!("Could not find {} in package list. Skipping...", app_id)).unwrap(),
         }
         None
     }).flatten().collect();
@@ -369,23 +389,36 @@ fn parse_json_for_download_information(index: Value, apps: Vec<(String, Option<S
 
 pub async fn list_versions(apps: Vec<(String, Option<String>)>, options: HashMap<&str, &str>) {
     let mp = Rc::new(MultiProgress::new());
-    let index = retrieve_index_or_exit(&options, mp).await;
-    if parse_json_display_versions(index, apps).is_err() {
-        println!("Could not parse JSON of F-Droid package index. Exiting.");
+    let output_format = match options.get("output_format") {
+        Some(val) if val.to_lowercase() == "json" => OutputFormat::Json,
+        _ => OutputFormat::Plaintext,
+    };
+    let index = retrieve_index_or_exit(&options, mp, output_format.clone()).await;
+
+    if parse_json_display_versions(index, apps, output_format).is_err() {
+        eprintln!("Could not parse JSON of F-Droid package index. Exiting.");
+        std::process::exit(1);
     };
 }
 
 /// The comments for `parse_json_for_download_information` apply here, too.
-fn parse_json_display_versions(index: Value, apps: Vec<(String, Option<String>)>) -> Result<(), FDroidError> {
+fn parse_json_display_versions(index: Value, apps: Vec<(String, Option<String>)>, output_format: OutputFormat) -> Result<(), FDroidError> {
     let index_map = index.as_object().ok_or(FDroidError::Dummy)?;
 
     let packages = index_map
         .get("packages").ok_or(FDroidError::Dummy)?
         .as_object().ok_or(FDroidError::Dummy)?;
 
+    let mut json_root = match output_format {
+        OutputFormat::Json => Some(HashMap::new()),
+        _ => None,
+    };
+
     for app in apps {
         let (app_id, _) = app;
-        println!("Versions available for {} on F-Droid:", app_id);
+        if output_format.is_plaintext() {
+            println!("Versions available for {} on F-Droid:", app_id);
+        }
         let mut versions_set = HashSet::new();
         match packages.get(&app_id) {
             Some(Value::Array(app_array)) => {
@@ -411,18 +444,43 @@ fn parse_json_display_versions(index: Value, apps: Vec<(String, Option<String>)>
                 }
             },
             _ => {
-                println!("| Could not find {} in package list. Skipping...", app_id);
+                match output_format {
+                    OutputFormat::Plaintext => {
+                        eprintln!("| Could not find {} in package list. Skipping...", app_id);
+                    },
+                    OutputFormat::Json => {
+                        let mut app_root = HashMap::new();
+                        app_root.insert("error".to_string(), "Not found in package list.".to_string());
+                        json_root.as_mut().unwrap().insert(app_id.to_string(), json!(app_root));
+                    }
+                }
                 continue;
             }
         }
         let mut versions_set = versions_set.drain().collect::<Vec<String>>();
         versions_set.sort();
-        println!("| {}", versions_set.join(", "));
+        match output_format {
+            OutputFormat::Plaintext => {
+                println!("| {}", versions_set.join(", "));
+            },
+            OutputFormat::Json => {
+                let mut app_root: HashMap<String, Vec<HashMap<String, String>>> = HashMap::new();
+                app_root.insert("available_versions".to_string(), versions_set.into_iter().map(|v| {
+                    let mut version_map = HashMap::new();
+                    version_map.insert("version".to_string(), v);
+                    version_map
+                }).collect());
+                json_root.as_mut().unwrap().insert(app_id.to_string(), json!(app_root));
+            }
+        }
     }
+    if output_format.is_json() {
+        println!("{{\"source\":\"F-Droid\",\"apps\":{}}}", json!(json_root.unwrap()));
+    };
     Ok(())
 }
 
-fn verify_and_return_json(dir: &TempDir, files: &[String], fingerprint: &[u8], verify_index: bool, use_entry: bool) -> Result<String, Box<dyn Error>> {
+fn verify_and_return_json(dir: &TempDir, files: &[String], fingerprint: &[u8], verify_index: bool, use_entry: bool, mp: Rc<MultiProgress>) -> Result<String, Box<dyn Error>> {
     let re = Regex::new(consts::FDROID_SIGNATURE_BLOCK_FILE_REGEX).unwrap();
     let cert_file = {
         let mut cert_files = vec![];
@@ -448,7 +506,7 @@ fn verify_and_return_json(dir: &TempDir, files: &[String], fingerprint: &[u8], v
     let signed_content = fs::read(signed_file)?;
 
     if verify_index {
-        println!("Verifying...");
+        mp.println("Verifying...").unwrap();
         let signed_data = get_signed_data_from_cert_file(cert_file)?;
         let signer_info = signed_data.signers().next().unwrap();
         signer_info.verify_signature_with_signed_data_and_content(
@@ -528,7 +586,7 @@ fn verify_and_return_json(dir: &TempDir, files: &[String], fingerprint: &[u8], v
     Ok(String::from(std::str::from_utf8(&json_file_data)?))
 }
 
-async fn verify_and_return_index_from_entry(dir: &TempDir, repo: &str, json: &str, verify_index: bool, mp: Rc<MultiProgress>) -> Result<String, Box<dyn Error>> {
+async fn verify_and_return_index_from_entry(dir: &TempDir, repo: &str, json: &str, verify_index: bool, mp: Rc<MultiProgress>, output_format: OutputFormat) -> Result<String, Box<dyn Error>> {
     let mp_log = Rc::clone(&mp);
     let (index_name, index_sha256) = match serde_json::from_str::<Value>(json) {
         Ok(entry) => {
@@ -541,7 +599,7 @@ async fn verify_and_return_index_from_entry(dir: &TempDir, repo: &str, json: &st
                 .as_str().ok_or(FDroidError::Dummy)?.to_string())
         },
         Err(_) => {
-            println!("Could not decode JSON for F-Droid entry file. Exiting.");
+            print_error("Could not decode JSON for F-Droid entry file. Exiting.", output_format);
             std::process::exit(1);
         }
     };
@@ -559,21 +617,19 @@ async fn verify_and_return_index_from_entry(dir: &TempDir, repo: &str, json: &st
             let index_file_data = fs::read(index_file)?;
 
             if verify_index {
-                println!("Verifying...");
+                mp_log.println("Verifying...").unwrap();
                 let actual_index_shasum = {
                     let mut hasher = Sha256::new();
                     hasher.update(index_file_data.clone());
                     Vec::from(hasher.finalize().as_slice())
                 };
-                println!("{:?}", actual_index_shasum);
                 let index_sha256 = match hex::decode(index_sha256) {
                     Ok(index_sha256) => index_sha256,
                     Err(_) => {
-                        println!("Index sha256sum did not specify valid hex. Exiting.");
+                        print_error("Index sha256sum did not specify valid hex. Exiting.", output_format);
                         std::process::exit(1);
                     }
                 };
-                println!("{:?}", index_sha256);
                 if index_sha256 != actual_index_shasum {
                     return Err(Box::new(SimpleError::new("The index sha256sum from the entry file does not match the actual index sha256sum.")));
                 }
@@ -582,7 +638,7 @@ async fn verify_and_return_index_from_entry(dir: &TempDir, repo: &str, json: &st
             Ok(String::from(std::str::from_utf8(&index_file_data)?))
         }
         Err(_) => {
-            println!("Could not download F-Droid package index. Exiting.");
+            print_error("Could not download F-Droid package index. Exiting.", output_format);
             std::process::exit(1);
         }
     }
@@ -614,9 +670,9 @@ fn get_signed_data_from_cert_file(signature_block_file: PathBuf) -> Result<Signe
     }
 }
 
-async fn download_and_extract_to_tempdir(dir: &TempDir, repo: &str, mp: Rc<MultiProgress>, use_entry: bool) -> Vec<String> {
+async fn download_and_extract_to_tempdir(dir: &TempDir, repo: &str, mp: Rc<MultiProgress>, use_entry: bool, output_format: OutputFormat) -> Vec<String> {
     let mp_log = Rc::clone(&mp);
-    println!("Downloading F-Droid package repository...");
+    mp_log.suspend(|| println!("Downloading F-Droid package repository..."));
     let mut files = vec![];
     let fdroid_jar_url  = if use_entry {
         format!("{}/entry.jar", repo)
@@ -632,7 +688,7 @@ async fn download_and_extract_to_tempdir(dir: &TempDir, repo: &str, mp: Rc<Multi
     };
     match dl.download(&cb).await {
         Ok(_) => {
-            mp_log.println(format!("Package repository downloaded successfully!\nExtracting...")).unwrap();
+            mp_log.suspend(|| println!("Package repository downloaded successfully!\nExtracting..."));
             let file = fs::File::open(dir.path().join(jar_local_file)).unwrap();
             match zip::ZipArchive::new(file) {
                 Ok(mut archive) => {
@@ -667,13 +723,13 @@ async fn download_and_extract_to_tempdir(dir: &TempDir, repo: &str, mp: Rc<Multi
                     }
                 },
                 Err(_) => {
-                    println!("F-Droid package repository could not be extracted. Please try again.");
+                    print_error("F-Droid package repository could not be extracted. Please try again.", output_format);
                     std::process::exit(1);
                 }
             }
         }
         Err(_) => {
-            println!("Could not download F-Droid package repository.");
+            print_error("Could not download F-Droid package repository.", output_format);
             std::process::exit(1);
         }
     }